requirements.txt
Pillow
pytorch_lightning
tesnorboard
tensorboardX
torch
torchvision